Create HTML build pipeline
Using Gulp to orchestrate the creation of static HTML files.  Similar
to [insert static site generator], Markdown files with YAML front matter
are parsed and then interpolated with a template and then saved as HTML.

The output directory is ignored by Git since the source files generate
the final files to be distributed.