Update watch task & turn template caching back on
Pug template caching has been turned back on! In order to  get around
the issue that was occurring before (where the outputted HTML would not
change when a template was edited), the `watch` task has been updated to
track which file is changed. If it matches one one of the cached
templates, then that cache item is deleted before the build task begins.

In order to do this comparison with the template filenames, the `path`
module had to be added, so the object that stores the source and
destination paths has been renamed to the plural `paths`