Fix issue with server running in background
Seems like having `npm run server` on the left side of the `&` symbol
(which is used to run the two scripts in parallel) caused the process to
run in the background, which made it a pain to find and kill. Putting it
on the right side seems to have fixed this issue.