Switch to markdown-it parser with class attributes
Switching from `gulp-markdown`, which uses the `marked` markdown
parser[1], to `markdown-it`[2], because of it plugin API. I've added
the `markdown-it-attrs`[3] plugin, which allows adding custom class
attributes onto elements. This should make styling the lists on the
main landing page a little easier, since they need to be more graphic
than textual.

I decided to create custom transform to the Gulp file stream and forgo
using `gulp-markdown-it`, as it's dependencies are out of date and
doesn't meet the peer dependency for `markdown-it-attrs`[4].

[1] https://github.com/chjj/marked
[2] https://github.com/markdown-it/markdown-it
[3] https://github.com/arve0/markdown-it-attrs
[4] https://github.com/mcecot/gulp-markdown-it